Add a --no-swtpm option to kola to support running kola tests without TPM
Architectures like s390x and pcc64le do not seem to have support for the virtual
TPM device in qemu. The kola tests fail to run today because it requires TPM support
by default.

With the --no-swtpm option specified, kola tests can be run on qemu images without TPM.
For the tests that require TPM such as rhcos.luks.tpm, they can be blacklisted in the config

+++ platform/machine/unprivqemu/machine.go
@@ -69,11 +69,13 @@ func (m *machine) Reboot() error {
 }
 
 func (m *machine) Destroy() {
-	if err := m.swtpm.Kill(); err != nil {
-		plog.Errorf("Error killing swtpm of %v: %v", m.ID(), err)
-	}
-	if err := os.RemoveAll(m.swtpmTmpd); err != nil {
-		plog.Errorf("Error removing swtpm dir: %v", err)
+	if m.swtpmTmpd != "" {
+		if err := m.swtpm.Kill(); err != nil {
+			plog.Errorf("Error killing swtpm of %v: %v", m.ID(), err)
+		}
+		if err := os.RemoveAll(m.swtpmTmpd); err != nil {
+			plog.Errorf("Error removing swtpm dir: %v", err)
+		}
 	}
 	if err := m.qemu.Kill(); err != nil {
 		plog.Errorf("Error killing instance %v: %v", m.ID(), err)
